chore: finish removing js form image.php
