+++ admin/controllers/images/views/show/script.js
@@ -114,7 +114,7 @@ function get_selected() {
 
 function get_selected_ids() {
 	thisarray=[];
-	selected = document.querySelectorAll('.all_images_image_container.active');
+	const selected = document.querySelectorAll('.all_images_image_container.active');
 	selected.forEach(selimage => {
 		thisarray.push(selimage.dataset.id);
 	});
@@ -122,7 +122,7 @@ function get_selected_ids() {
 }
 
 function rename_image() {
-	selected = get_selected();
+	const selected = get_selected();
 	if (selected.length<1) {
 		alert('Select an image');
 	}
@@ -144,7 +144,7 @@ function rename_image() {
 }
 
 function crop_image() {
-	selected = get_selected();
+	const selected = get_selected();
 	if (selected.length<1) {
 		alert('Select an image');
 	}
@@ -290,4 +290,20 @@ function delete_items() {
 		alert('No images selected');
 	}
 }
+
+document.querySelector("[data-clickaction='clear_selection']").addEventListener("click", ()=>{
+	clear_selection();
+});
+document.querySelector("[data-clickaction='rename_image']").addEventListener("click", ()=>{
+	rename_image();
+});
+document.querySelector("[data-clickaction='crop_image']").addEventListener("click", ()=>{
+	crop_image();
+});
+document.querySelector("[data-clickaction='clear_tags']").addEventListener("click", ()=>{
+	clear_tags();
+});
+document.querySelector("[data-clickaction='delete_items']").addEventListener("click", ()=>{
+	delete_items();
+});
